Modify counterbores and screw holes in picam2_cutout picamera_2_camera_mount and lens_spacer
picam2_cutout had the hole for self tapping built in.
Remove and place as 'if-else' in the picamera_2_camera_mount module
where it selects counterbores or not.
Countebores redefined as r=1.15 not r=1 for No2 self-tap instead of M2

lens_spacer has the counterbores re-asserted at the end as there was an overhang
shelf protruduing into the counterbore from the convex hull making up the spacer.