Cutout for the illumination cable
I've made a hole at the front of the Z cable tidy for the
illumination cable.
Doing this meant figuring out a bunch of the transforms.
I have added comments as I went.
I also needed to get the rendering working on my
computer, and have added a README in case it helps
others in the future.