Added screws to retain small gear
Removed bevels. Re-did the counterbore for the round section of the shaft
so gear sits 1/2 mm above motor for the orientation we use in instructions.
Gear height increased slightly to compensate.

Tightened shaft hole on gear as it was very very loose on Prusa i3 Mk3.
With splodgyer printers there may be a need to need press fit the gear
with a clamp or vice. Tighter test prints (down to flat_shaft_w=3) on a
Prusa i3 Mk3 press fitted well.

The self-tapping screws go in for all sizes as they cut into the plastic
this is intended to tighten the plastic keeping it well fitted in tension
against the shaft. Holes are counterbored so that screwhead sits flush with
top of gear.