First non-working Kelvin mount
This adds a plate and a mount with balls and dowel pins, to make a
Kelvin mount for the objective.

I've modified the optics library to make the RMS thread easier to borrow
but should not have changed anything else.

Currently the balls/pins don't protrude high enough so the parts don't
mate correctly