Shift, rather than duplicate, the mounting bolt
I've added in a parameter that lets us shift where the mounting
bolt sits on the optics module.  This allows me to make the larger
mount fit into the upright, with the fewest changes to the code.

I'm not sure if this is the cleanest way to do it, but it avoids
propagating loads of extra arguments,
which is why we started using dictionaries in the first place...