Remove explicit arguments for condenser()
The condenser is rendered using `condenser()` but the STL specified
the argument values.  This worked
when the two files (`illumination.scad` and `condenser.scad`) were
in sync, but goes wrong if one of them changes.  I have used defaults
in `condenser.scad` and added a comment explaining why this is done.