Factor out arguments to condenser()
The arguments to `condenser()` were being duplicated, and this
caused inconsistencies when they changed in one place.
In the long term it would be lovely for this to be passed around as a
fully parametric dictionary, similar to the optics module parameters.
For now, I have removed the arguments and created functions
to make their (now fixed) values accessible.

This fixes the problem - they won't go out of sync - and should also
make it easier to search-and-replace in the future, to make it
properly parametric.