Handle condenser_base_r consistently
The position of the condenser mounting screws depends on the
lens diameter, which is now fixed.
Rather than bake the function in everywhere, I've continued passing
around the condenser base_r, always
calculating it with the relevant function, using `condenser_lens_d()`
as its argument.

This means that, in the future, only `condenser_lens_d()` needs to be
replaced with a changeable parameter: the rest is all calculated
from that value.