Flip the nut entry slot on the condenser
The locking nut in the condenser is inserted through an angled slot.
The slope is important for strength, but meant that, if the thumbscrew
is removed, the nut will fall out.
This commit flips the slot so that the nut will be retained by gravity,
when the condenser is used as intended in the inverted microscope.

A comment has been added to explain why the slot must be sloping.

I've updated the render of the nut entry, so that it still enters from
the right direction.