Improve bridging and aesthetics in cable tidies
The main point of this commit is to improve the bridging in the cable
tidies.  The void for the cable is divided into two parts with different
roof heights, so the bridging is only ever along one axis - that should
slice much better.

I have also rounded the top edges which should result in a more
pleasing shape.