Specify minimum gitbuilding version
Something went wrong (PyPI outage) and we used the wrong
version of gitbuilding.  We rely on features introduced in gitbuilding
0.11.0, so I now specify we need
gitbuilding>=0.11.0.

Ordinarily this should not be a problem - gitbuilding should
always install at least 0.11.0.  However, it seems reasonable to
specify a minimum version given that we know we require one.