Make slice_gap_closing_radius universal
I've simplified the instructions for printing, and suggested
changing the slice gap closing radius for all parts.  I don't
think this should cause any problems.

I've also added this to the print settings in the `test your printer`
page, and tried to improve the text in a couple of places.

Closes #235