Add instructions to build 35mm parfocal optics
This does not build the modules by default, because it
would add ~18 minutes to the build time, which is
already longer than I'd like. Instead, I have made it
simple to compile them by passing `-D` options to
`rms_optics_module.scad`, and described this in
the instructions.

In due course, an external repository that builds these
would be a nice solution.