Don't hull the beamsplitter with the camera
The optics module adds space for the beamsplitter by
taking a convex hull of several objects.  I think this should
only be the beamsplitter casing and the top and bottom
of the optics module cylinder.

Previously it also included the camera module - I think
doing this is likely to make a mess for large cameras.
I believe simply removing the camera module should
revert to the intended behaviour.

I have not tested the outputs of this commit - I am
pushing so I'll see them in CI.