Refactor optics module
I've tidied up the optics module into a number of
submodules in an effort to simplify the structure.
This also sorts out the "hull"ing by adding in a module
that calculates the bottom of the fluorescence part.
That eliminates the problem of large camera mounts
making the casing engulf the fitting wedge.