New light trap based on polyhedron()
Based on the success of the revised rms_thread_cutter,
I have implemented a polyhedron() based solution to
the light trap cylinder.

The hope is to improve reliability, i.e. reduce mesh errors
while also improving speed.  If this is successful, I'll
make a revised version for `lighttrap_sqylinder` as well.
In fact, I can unify the two functions I think, because
the overhead of adding in the square is minimal.