Build CSG files
Currently we build the STLs directly from OpenSCAD
source.  It's hard to determine which STLs have changed,
because there are a lot of interdependencies where we
may not actually have modified the functions that are
used.

Also, OpenSCAD builds STLs non-deterministically - so
often an output STL will change even if the geometry is
nominally identical.

If we build (and cache, or at least hash) the CSG files, we
could:
* Only build STLs that have actually changed
* Maintain a list of STLs that have changed (or not) from
  version to version - which would be useful.

This commit modifies the build script to produce those
CSG files, and it takes under a minute for the whole project.

NB I have worked around an OpenSCAD bug (I think)
with relative paths for CSG export.