Add CSG fixup step
CSG files can generate warnings when compiled with OpenSCAD
in particular every instance of `import` results in a warning about
the `timestamp` argument.

This commit adds a Python processing step to remove the
timestamps. This solves two problems:
1. The CSG file is deterministic (no dependence on timestamps of
input files)
2. OpenSCAD no longer fails because of the warning (we have
`--hardwarnings` enabled in the OpenSCAD rule, and it would be
best not simply to turn this off)