Copy included DXF files to output directory
Intermediate CSG files in builds have relative import paths.  This means
they fail if the CSG file is not in the same place as the SCAD.
For now, I'll copy the imported files to the build directory - it may be
desirable in the future to put the CSG files in the source directory,
though this seems messy to me.
Maybe a third directory (for intermediate CSG files) would be the right
answer?