First working cached CSG compiler
I've now added a wrapper script that will check the cache before
compiling a CSG -> STL, and leave the existing STL if it matches the
stored hash.

This should work correctly if it is run against a build directory that
contains dependency_hashes.yaml
and a matching set of STL files.
If there are STL files that don't match the hash file, it will
attempt to rebuild them.

I have not yet added rebuilding
the hash file into the build script,
that would be a good idea!