Refactor and deduplicate optics module
I now use the main RMS optics module, without any copypasta, and have
split out the part that's added on to the optics module to support the
Kelvin mount plate.

I think this should now be in a much better state of readiness for
merge, we just need to figure out the spacer height!