Update license to CERN-OHL-S
This is a long overdue and uncontroversial
switch from CERN-OHL-1.2 to CERN-OHL-S. It is an update
but not a change of the terms in
principle.

All files that specified a license only stated
"CERN Open Hardware License" so
I haven't changed this.

I note that "actuator_tension_band" specified
v1.2, this may need to be removed
or updated - after checking with
the contributor in question.

I've also updated some contact
details and version numbers.