Shrink self tapping holes to fit no. 2
The holes are currently intended to "self tap" M2.5 screws. I've
reduced this to fit M2 screws, which sould work better with
the self tappers used elsewhere.

This probably requires counterboring the plate that screws on
to mount the illumination.

We should swap this to properly use the module for a
no.2 self tap hole before merging.