Update instructions on firmware
I've moved flashing firmware out of the instructions as
most users won't need to do this unless they are already
using the work-around page.

We should ensure the sangaboard page describes how to upload the firmware, and links to a binary.

We should ensure the workaround page does the same for both RP2040 and Nano based boards.