Conical wedge for camera posts. Other lower level changes ...
In making the camera posts conical the faces of the fitting wedge
were no longer the outer shape of the hulled piece.

Rather than fix this ad-hoc per camera mount, new code was added
to optionally cut the faces of the fitting wedge in
`fitting_wedge_cutout()` and `objective_fitting_cutout()`

Some further refactoring added to the code that makes the posts for
the camera platforms to reduce the pyramid of doom.

Closes #340