Reinstate counterbore for locking dovetail if not tapered
Locking dovetails are tapered for all parts in the main microscope.
But some additions such as the LED array holder the block isn't
tapered and so the face the screw/thumbscrew meets isn't perpendicular
to the shaft. Pre !222 there was a counterbore for all dovetails. This
MR reinstates the counterbore only if the block is not tapered.

Closes #213