Adjust paths of import statements
import statements of the .csg files break as they are no longer in the
location of the .scad files. Read the original locations from the dependency
files and replace the path.

This worked so far because the build.py copies the imported logo dxf
files to docs/models/libs. This might no longer be needed. Using the files
from the openscad/ subdir also revents the stl to be rebuild every time.

Code is a bit complicated as it does not rely on there being only one file
with the same basename being imported. In case there are multiple we use the
one with the longer matching path.

This change is needed for the Delta to use the build_system.