Generate dependency files for the fixed csg files
The rule for the .fixed.csg files has depfile="$out.d". This make ninja
assume this file is actually created by fix_csg.py. The file not being
there results in ninja building the file and the stl every time.

Yes, the caching via cached_csg_compiler's dependency_hashes.yaml does
prevent that for prebuild stls as long as no dependencies are changed.
But Ninja is much smarter with this and makes things a lot easier during
local development.