Implement dynamic caching
Place .sha1 files beside the fixed.csg files and copy it when
building the stl from it. Use these in cached_csg_compiler to
skip building stls when the fixed.csg file is unchanged.

This is how the caching is probably supposed to work from the start.
We can probably get rid of the dependency_hashes.yaml file and most
code related to it. Need to package the sha1 files in the CI first.