Add upright optics modules
As the camera is at a 45° angle to the mounting dovetail flipping it
upside down turns is 90°. To keep the camera alligned with the X and Y
axis of the stage the upright microscope needs optics modules that have
the camera turned 90°. This also points the ribbon cable towards the Y
actuator which makes cable management easier.

Unfortunately the b0196 camera does not fit in this orientation and is
turned the other way round.

Also add comments to all the top level scad files.